reorganise test suits
