+++ tests/unit_tests/bin_manager_test.py
