+++ tests/unit_tests/diamonds_test.py
