+++ tests/unit_tests/main_binette_test.py
@@ -15,7 +15,7 @@ from binette.main import (
     manage_protein_alignement,
     parse_input_files,
 )
-from tests.bin_manager_test import (  # noqa: F401
+from tests.unit_tests.bin_manager_test import (  # noqa: F401
     create_temp_bin_directories,
     create_temp_bin_files,
 )
